Adjusted offset so human is correctly positioned relative to origin. Disable scaling via allegro for now - too slow! Camera boundary is now defined by terrain size. Now profiling GUI draw time. World draw is now being profiled. We're now drawing terrain again. Temporarily displaying Z order of sprites for debugging. Temporarily displaying entity origin and bounds for debugging. Added toggles for FPS and profiling display.

+++ src/game/Terrain.h
@@ -58,7 +58,7 @@ namespace ct
 			TerrainTileCorner corners[ 2 ];
 			float height[ 4 ];
 
-			void Draw( const Camera &camera, float offsetX, float offsetY );
+			void Draw( const Camera &camera, int offsetX, int offsetY );
 		};
 
 	private:
